requirements.txt
Flask
Flask-SQLAlchemy
Flask-APScheduler
python-dotenv
google-genai
pydantic
ccxt
newsapi-python
praw
requests
gunicorn
